requirements.txt
numpy>=1.19.5
pandas>=1.1.5
matplotlib>=3.4.3
scikit-learn>=1.0.2
tensorflow>=2.8.0
opencv-python>=4.5.5
streamlit>=1.8.1
pillow>=8.4.0